Added Talk Icon
Fixed a swimming bug where the player would only swim fast if it was selected in the hierarchy. Added a talk icon that now disables itself when text is playing. Fixed the bug where the player would automatically jump when the dialogue ended. This was because the dialogue continue button was the same as the jump button.